src/main/java/com/github/cc3002/finalreality/model/character/player/PlayerCharacter.java
package com.github.cc3002.finalreality.model.character.player;

import com.github.cc3002.finalreality.model.character.AbstractCharacter;
import com.github.cc3002.finalreality.model.character.ICharacter;
import com.github.cc3002.finalreality.model.weapon.IWeapon;
import java.util.concurrent.BlockingQueue;
import java.util.concurrent.Executors;
import java.util.concurrent.TimeUnit;

import org.jetbrains.annotations.NotNull;

/**
 * A class that holds all the information of a single character of the game.
 *
 * @author Ignacio Slater Muñoz.
 * @author Diego Ruiz R.
 */
public class PlayerCharacter extends AbstractCharacter {
  protected IWeapon equippedWeapon;

  public PlayerCharacter(@NotNull BlockingQueue<ICharacter> turnsQueue, @NotNull String name)
  {
    super(turnsQueue, name);
    equippedWeapon=null;
  }

  /**
   * Returns this character's equippedWeapon
   *
   */
  public IWeapon getEquippedWeapon() {return this.equippedWeapon;}


  /**
   * Sets a scheduled executor to make this character (thread) wait for {@code speed / 10}
   * seconds before adding the character to the queue.
   */
  public void waitTurn() {
    scheduledExecutor = Executors.newSingleThreadScheduledExecutor();
      scheduledExecutor
              .schedule(this::addToQueue,getEquippedWeapon().getWeight() / 10, TimeUnit.SECONDS);

  }
}